Add files via upload
In part 1, user enters two integers and a operation(add or multiplicate). It adds or multiplicates all numbers between user-determined integers and prints the result.
Part 2 gets an integer from user, shows all prime and non-prime numbers(and shows the minimum integer that divide them) from 1 to the our integer.
Part 1 writes its results to a file.
Part 3 reads that file and prints what's inside.
Part 4 sorts integers in the file and prints the new sequence.